requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
attrs==21.2.0
awscli==1.20.26
black==21.7b0
boto3==1.18.25
botocore==1.21.26
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.3
decorator==5.1.0
docutils==0.15.2
firebase-admin==5.0.2
Flask==2.0.1
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
flask-restx==0.5.0
Flask-SQLAlchemy==2.5.1
google-api-core==1.31.2
google-api-python-client==2.18.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud-core==2.0.0
google-cloud-firestore==2.3.1
google-cloud-storage==1.42.0
google-crc32c==1.1.3
google-resumable-media==2.0.0
googleapis-common-protos==1.53.0
greenlet==1.1.1
grpcio==1.39.0
gunicorn==20.1.0
httplib2==0.19.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
jsonschema==3.2.0
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
msgpack==1.0.2
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
pbr==5.6.0
proto-plus==1.19.0
protobuf==3.17.3
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.3
requests==2.26.0
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.23
sqlalchemy-migrate==0.13.0
sqlparse==0.4.2
stripe==2.60.0
Tempita==0.5.2
tomli==1.2.1
uritemplate==3.0.1
urllib3==1.26.6
Werkzeug==2.0.1